+++ src/App.js
@@ -5,11 +5,16 @@ import { Route, Routes } from 'react-router-dom';
 import Join from './JoinLogin/Join';
 import { useState } from 'react';
 import EmailJoin from './JoinLogin/EmailJoin';
+import JoinTemsPage from './JoinLogin/JoinTemsPage';
+
+
 
 function App() {
 
   return (
     <div className="App">
+    
+    {/* {false && <Header/>} */}
     <Header/>
     <Routes>
       {/* <Route path = '/login' element={<Login userInfo = {userInfo} 
@@ -17,6 +22,7 @@ function App() {
                                                setImsiName = {setImsiName}/>}/> */}
       <Route path='/join' element={<Join/>}/>
       <Route path='/emailJoin' element={<EmailJoin/>}/>
+      <Route path='/tems' element={<JoinTemsPage/>}/>
 
     </Routes>
